requirements.txt
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
commonmark==0.9.1
docutils==0.19
idna==3.4
importlib-metadata==5.0.0
jaraco.classes==3.2.3
keyring==23.9.3
more-itertools==8.14.0
numpy
pandas==1.5.1
pkginfo==1.8.3
Pygments==2.13.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pywin32-ctypes==0.2.0
readme-renderer==37.2
requests==2.28.1
requests-toolbelt==0.10.0
rfc3986==2.0.0
rich==12.6.0
six==1.16.0
soupsieve==2.3.2.post1
twine==4.0.1
urllib3==1.26.12
webencodings==0.5.1
zipp==3.9.0
fuzzywuzzy
python-Levenshtein